Updated SnapshotWriter to include conserved quantities:   totalMass,   totalEnergy,   totalMomentumX,   totalMomentumY in .npy zip.
Originally tried to put these in header of .npz file, but python3 was complaining about structure of header. Decided to just save the data as a separate numpy array in the same .npz file

To do so, made these quantities Public in the Fluid class of SimpleKHI.cpp and adjusted the call to save_npz_snapshot() accordingly.