Mostly modified plotting routines to make pngs for move of conserved quantities
Note that the fluid.cpp was modified to include time (t) variable in the saving to snapshot call. The SnapshotWritter.hpp was modified accordingly. However, it should be tested further to ensure it works.

Similarly, the ConservedQuantityPlotter.py assumes the time variable is part of the saved snapshot. If not, then either perform quick-and-dirty solution in the .py file (create a local time array from the number of snapshots and the timestep) **or** change the fluid.cpp and SnapshotWritter.hpp appropriately.

Directory structure for images was modified slightly.
Could use work to make it less of a mess.